separate application logic from silo deployment, and make the silo deployment a two-silo deployment with real membership logic
