--- BasicOrleansApp/Program.cs
@@ -1,29 +1,0 @@
-﻿using Microsoft.Extensions.DependencyInjection;
-using Microsoft.Extensions.Hosting;
-using Orleans;
-using Orleans.Hosting;
-using Silo;
-
-// Configure the host
-using var host = new HostBuilder()
-    .UseOrleans(builder => builder.UseLocalhostClustering())
-    .Build();
-
-// Start the host
-await host.StartAsync();
-
-// Get the grain factory
-var grainFactory = host.Services.GetRequiredService<IGrainFactory>();
-
-// Get a reference to the HelloGrain grain with the key "friend"
-var friend = grainFactory.GetGrain<IHelloGrain>("friend");
-
-// Call the grain and print the result to the console
-var result = await friend.SayHello("Good morning!");
-Console.WriteLine("\n\n{0}\n\n", result);
-
-Console.WriteLine("Orleans is running.\nPress Enter to terminate...");
-Console.ReadLine();
-Console.WriteLine("Orleans is stopping...");
-
-await host.StopAsync();
