+++ ServerlessOrleansHackathon.sln
@@ -5,7 +5,9 @@ VisualStudioVersion = 17.3.32929.385
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "ConnectionTest", "ConnectionTest\ConnectionTest.csproj", "{F79EC084-7BC2-4240-8443-7AD0BE4A115D}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Silo", "BasicOrleansApp\Silo.csproj", "{011C55A8-BDBA-48B2-ADF1-21CCA9F0ABF1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TwoLocalSilos", "TwoLocalSilos\TwoLocalSilos.csproj", "{011C55A8-BDBA-48B2-ADF1-21CCA9F0ABF1}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Application", "Application\Application.csproj", "{70525322-152B-4753-B2C6-DDA70DB97FE9}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -21,6 +23,10 @@ Global
 		{011C55A8-BDBA-48B2-ADF1-21CCA9F0ABF1}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{011C55A8-BDBA-48B2-ADF1-21CCA9F0ABF1}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{011C55A8-BDBA-48B2-ADF1-21CCA9F0ABF1}.Release|Any CPU.Build.0 = Release|Any CPU
+		{70525322-152B-4753-B2C6-DDA70DB97FE9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{70525322-152B-4753-B2C6-DDA70DB97FE9}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{70525322-152B-4753-B2C6-DDA70DB97FE9}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{70525322-152B-4753-B2C6-DDA70DB97FE9}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
